Create README.md
This is Sherin Naguib from E solutions company in Egypt. we are under process of launching new eticketing system that provides tickets for amusements occasions in Egypt such as events and concerts. therefore, we would like to check if we can collaborate with you to create software as a white label partner. we can discuss it further through the email address: sherin.naguib@dominick-esolutions.com or phone no. 0020100534713